refactor: atualização dos tipos de dados na resposta e requisição dos DTOs, melhoria na lógica de cálculo de produção.

+++ backend/src/main/java/com/factory_management/entities/Product.java
@@ -24,9 +24,9 @@ public class Product {
   private Integer amount;
 
   @Column(nullable = false)
-  private float price;
+  private int price;
 
-  public Product(String name, int amount, float price) {
+  public Product(String name, int amount, int price) {
     this.name = name;
     this.amount = amount;
     this.price = price;
